requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
frozenlist==1.3.0
future==0.18.2
idna==3.3
line-bot-sdk==2.2.1
multidict==6.0.2
python-decouple==3.6
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
yarl==1.7.2
